allows disabling logging to file
so the cheater does not know exactly why he was flagged!

TODO: still 1 log appearing at start, due to TLSCallback thread created at start, and log in UnmanagedGlobals::AddThread, in main.cpp. So a file with 1 line is still created.
TODO: in the same vein, create a static "Logger::filename" variable, so instead of writing "UltimateAnticheat.log" in every log call, we define it only once.
(and if by default we set it to "", it will fail the very first log call)