WIP documentation (do not merge).
Hi, it's getting late for me, so I only started a skeleton of a documentation, more important is asking for your approval since I try to explain your stuff , from your comments and some quick google searches.

Instead of creating a documentation folder, Ideally a Github wiki would be nice.

However, it's not possible to do pull requests on wiki pages, you have to either :
- edit the wiki pages yourself
- allow users as collaborators (risky)
- allow public editing of the wiki pages (by anyone with a github account)

so not ideal, we can start with a documentation folder if you want.