prevent MonitorProcessCreation thread crash on shutdown
when cleanly shutdowning the app, it crashed as the MonitorProcessCreation hanged.

replace the process enumerator hanging indefinitely (WBEM_INFINITE) with  an immediate polling (WBEM_NO_WAIT) that loops every 0,1 sec. (it' works with a queue, so we don't miss processes)

added mutex to prevent MonitorProcessCreation crash. (if shutdowning during process scanning, for example)