Put HasSignature in thread with timeout
Some calls to WinVerifyTrust are indefinitely hanging, so putting them in threads with a 15s timeout, and if failing, log it as an error but silently fail by returning `true` (saying the DLL is signed, even if we couldn't check)

that's not great, but I don't know what else to do, as otherwise one single call can just block a whole thread.

I created a [Stackoverflow issue](https://stackoverflow.com/questions/79446099/winverifytrust-hangs-or-crashes-with-certain-dlls) on the subject, but for now people seem to be nitpicking on it...
I think I don't have the crash I spoke about on the issue, I think it was due to a mistake in my thread argument passing.
Still, needs testing on multiple conditions.