fix infinite hangups and crashes at process stop
Now  that I killed some infinite Signature checks, I noticed that other methods hanged indefinitely and crashed if we tried to stop them!

Added std::mutex and conditions to gracefully stop current tasks without waiting too long, and without having a crash because a thread was deleted while running

fix infinite hangup for `WaitForMultipleObjects(...INFINITE)` when we were already in a loop and its a method with a queue.

comment call to  `GetUnsignedDrivers` because for now it's not really used, and it should be moved to its own thread, as it takes way too long and hangs the main thread.